% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributedLMM.R
\docType{package}
\name{DistributedLMM-package}
\alias{DistributedLMM}
\alias{DistributedLMM-package}
\title{DistributedLMM: Predicting Length of Hopsital Stay for COVID-19}
\description{
Predicting Length of Hopsital Stay for COVID-19.
}
\keyword{internal}
